Silenced near all /W3 MSVC warnings across project Should make the default Visual Studio warning flags much less noisy Nearly all warnings were pure noise, but the effort did yield 1 bug

+++ include/dart/api.tcc
@@ -198,7 +198,7 @@ namespace dart {
   template <class Number>
   basic_number<Number>::operator int64_t() const noexcept {
     if (is_integer()) return integer();
-    else return decimal();
+    else return static_cast<int64_t>(decimal());
   }
 
   template <class Number>
